# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Memora API
  description: |
    REST API for the Memora flashcard study app.

    **Base URL:** `/api` (same origin on Vercel, or `http://localhost:5001/api` in local dev).

    **Auth:** Send `Authorization: Bearer <JWT>` on protected routes.

    **Roles:** `USER`, `ADMIN`, `SUPER_ADMIN`. Staff (`ADMIN`, `SUPER_ADMIN`) see all subjects and can use admin routes.

    **Rate limits:** Auth-related POST routes (login, register, guest, Google, password reset, verify, close account) — 40 requests per 15 minutes per IP. Progress POST — 150 requests per minute per IP. Contact POST — 5 requests per hour per IP.

    **Account lifecycle:** Users are **deactivated** (not hard-deleted). Learners may **close** their own account. Super admins can **reactivate** deactivated users and **verify email** manually.

    **Deployment:** Production uses one Vercel serverless handler (`api/index.js`). Fast paths use `pg` and avoid loading Express when possible; remaining traffic falls back to Express + Prisma.

    **CI:** Pushes to `main` run unit/smoke tests via GitHub Actions (`npm run ci`). Vercel builds use `scripts/build-vercel.sh` (tests skipped on Vercel; Prisma Linux engine only).

    **Vercel fast paths** (representative): health, auth (login/register/guest/Google/verify/password), subjects catalog, progress POST, admin reports, many admin list/import routes. Subject/topic/flashcard mutations and some admin exports may use Express on first hit.

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [USER, ADMIN, SUPER_ADMIN]
        createdAt: { type: string, format: date-time }
        emailVerified: { type: boolean }
        hasPassword: { type: boolean }
        hasGoogle: { type: boolean }
        isGuest: { type: boolean }
        active:
          type: boolean
          description: Present on admin user list; false when deactivated
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
    MeResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
...
    Message:
      type: object
      properties:
        message: { type: string }
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, subjectIds]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                subjectIds:
                  type: array
                  minItems: 1
                  items: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email with token from link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email (authenticated)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/guest:
    post:
      tags: [Auth]
      summary: Start a guest session (no email/password)
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/upgrade-guest:
    post:
      tags: [Auth]
      summary: Convert guest account to registered user
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
        '409':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/google:
    post:
      tags: [Auth]
      summary: Sign in or register with Google ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential: { type: string }
                subjectIds:
                  type: array
                  items: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/AuthResponse' }
                  - type: object
                    properties:
                      needsSubjectSetup: { type: boolean }
        '400':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /auth/close-account:
    post:
      tags: [Auth]
      summary: Close own account (self-deactivate)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: Required for password-based accounts
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400':
        '403':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '401':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
                confirmPassword: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400':
        '401':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...